28_python/poetry_pure/pyproject.toml
[tool.poetry]
name = "20-poetry"
version = "0.1.0"
description = ""
authors = ["Chris Guest <chrisguest75@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0.1"
python-json-logger = "^2.0.7"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.0.2"
flake8 = "^7.0.0"
flake8-bugbear = "^24.2.6"
flake8-2020 = "^1.8.1"
black = "^24.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[scripts]
"start" = "python main.py"
"start:test" = "python main.py --test"
test = "pytest"
lint = "flake8"